Refactor token request schemas
Remove oneOf to support creation and token validation in different
requests. Code generators seem to have problems to properly support this
in combination with form parameters.

Instead, just send all relevant parameters in the same request.
Guacamole will only issue a new token if the old one in the payload is
invalid.